merge

+++ db/tables/users.js
@@ -5,7 +5,7 @@ const sql = `CREATE TABLE IF NOT EXISTS users (
   email VARCHAR(255) NOT NULL, 
   username VARCHAR(255) NOT NULL, 
   password VARCHAR(255) NOT NULL,    
-  total_reward INTEGER DEFAULT 0, 
+  total_reward FLOAT DEFAULT 0, 
   num_of_questions INTEGER DEFAULT 0, 
   choose_answers INTEGER DEFAULT 0, 
   num_of_answers INTEGER DEFAULT 0,
